#11 refactor: 디렉토리 구조 변경 및 코드 컨벤션 수정
