+++ src/App.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from "styled-components";
 import "./App.css";
 import GlobalStyles from "@styles/GlobalStyles";
 import designSystem from "@styles/designSystem";
-import NavBar from "@components/common/NavBar/NavBar";
+import NavBar from "@components/NavBar/NavBar";
 
 function App() {
   return (
